fix: concat dtype preservation and filter_tensor typed zero
- concat: only convert empty arrays to match non-empty dtype (was forcing all int to int32)
- filter_tensor: use tf.cast(0, value.dtype) instead of bare 0

Fixes sobol_generating_matrices returning int32 when int64 requested.

+1 test (test_sobol_generating_matrices_with_dtype)

+++ tf_quant_finance/math/qmc/utils.py
@@ -160,4 +160,4 @@ def filter_tensor(value: types.IntTensor, bit_mask: types.IntTensor,
       tf.bitwise.bitwise_and(
           tf.bitwise.right_shift(bit_mask, bit_index), tf.cast(1, value.dtype)))
 
-  return tf.where(is_bit_set, value, 0)
+  return tf.where(is_bit_set, value, tf.cast(0, value.dtype))
